pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "by-flask-starter"
version = "0.1.0"
description = "Minimal production-ready Flask starter"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Bogdan Yatsenko" }]

dependencies = [
    "flask>=3.0,<4.0",
    "python-dotenv>=1.0,<2.0",
    "gunicorn>=21,<22",
    "itsdangerous>=2.2,<3.0",
    "click>=8.1,<9.0",
    "werkzeug>=3.0,<4.0",
    "jinja2>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<6",
    "flake8>=7,<8",
    "black>=24,<25",
]

[tool.setuptools.packages.find]
where = ["src"]
